requirements.txt
asgiref==3.6.0
attrs==22.2.0
colorama==0.4.6
Django==4.1.7
django-birthday==0.1.4
django-bootstrap5==22.2
django-filter==22.1
exceptiongroup==1.1.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pytest==7.2.1
pytest-django==4.5.2
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
